hamad/3월 2주차/2문제 * 코드오류수정
