--- hamad/week2/src/Question1920.java
@@ -1,36 +1,0 @@
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.util.StringTokenizer;
-
-public class Question1920 {
-
-    public static void main(String[] args) {
-
-    }
-
-    static class Reader {
-
-        BufferedReader br;
-        StringTokenizer st;
-
-        public Reader() {
-            this.br = new BufferedReader(new InputStreamReader(System.in));
-        }
-
-        String next() {
-            while (st == null || !st.hasMoreElements()) {
-                try {
-                    st = new StringTokenizer(br.readLine());
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-            }
-            return st.nextToken();
-        }
-
-        int nextInt() {
-            return Integer.parseInt(next());
-        }
-    }
-}
